Outsource frontend auth checker
